+++ src/main/java/com/atlauncher/data/PackVersion.java
@@ -40,7 +40,9 @@ public class PackVersion {
 
     public String toString() {
         if (this.minecraftVersion == null || (this.minecraftVersion.id.equalsIgnoreCase(this.version)
-                && this.minecraftVersion.type != VersionManifestVersionType.SNAPSHOT)) {
+                && this.minecraftVersion.type != VersionManifestVersionType.SNAPSHOT
+                && this.minecraftVersion.type != VersionManifestVersionType.EXPERIMENTAL_SNAPSHOT
+                && this.minecraftVersion.type != VersionManifestVersionType.PENDING)) {
             return this.version;
         }
 
@@ -49,8 +51,19 @@ public class PackVersion {
             return this.version + " (Snapshot)";
         }
 
+        if (this.minecraftVersion.id.equalsIgnoreCase(this.version)
+                && (this.minecraftVersion.type == VersionManifestVersionType.EXPERIMENTAL_SNAPSHOT
+                        || this.minecraftVersion.type == VersionManifestVersionType.PENDING)) {
+            return this.version + " (Experiment)";
+        }
+
         if (this.minecraftVersion.type == VersionManifestVersionType.SNAPSHOT) {
             return this.version + " (" + this.minecraftVersion.id + ")" + " (Snapshot)";
+        }
+
+        if (this.minecraftVersion.type == VersionManifestVersionType.EXPERIMENTAL_SNAPSHOT
+                || this.minecraftVersion.type == VersionManifestVersionType.PENDING) {
+            return this.version + " (" + this.minecraftVersion.id + ")" + " (Experiment)";
         }
 
         return this.version + " (" + this.minecraftVersion.id + ")";