+++ src/main/java/com/atlauncher/data/minecraft/VersionManifestVersionType.java
@@ -22,6 +22,28 @@ import com.google.gson.annotations.SerializedName;
 
 @Json
 public enum VersionManifestVersionType {
+    @SerializedName("experimental_snapshot")
+    EXPERIMENTAL_SNAPSHOT {
+        public String getValue() {
+            return "experimental_snapshot";
+        }
+
+        public String toString() {
+            return "Experiment";
+        }
+    },
+
+    @SerializedName("pending")
+    PENDING {
+        public String getValue() {
+            return "pending";
+        }
+
+        public String toString() {
+            return "Experiment";
+        }
+    },
+
     @SerializedName("snapshot")
     SNAPSHOT {
         public String getValue() {